backend/sensor-hub/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
pyyaml==6.0.1
python-dateutil==2.8.2
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3
requests==2.31.0
pyserial==3.5
pynmea2==1.19.0

# Hardware interface libraries
RPi.GPIO==0.7.1
smbus2==0.4.3
spidev==3.6
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-mpu6050==1.3.4
adafruit-circuitpython-ssd1306==2.12.21
adafruit-circuitpython-cc1101==1.0.0

# Data processing
scipy==1.11.4
scikit-learn==1.3.2

# System monitoring
psutil==5.9.6
GPIO==0.1.0
